Increase the recommended min_free_kbytes value
Upon investigation, it was found that the recommended value for
min_free_kbytes was not reducing the number of fragmentation-related
events that were occuring in the page allocator enough.

This patch increases min_free_kbytes such that at least two pageblocks
are free for MIGRATE_RESERVE and 3 pageblocks free per migratetype that
is commonly used in the page allocator. This significantly reduces the
number of fragmentation-related events that occur.

However, it should be noted that to really get the fragmentation events
down, the upstream kernel needs to have the patch "page-allocator: Limit
the number of MIGRATE_RESERVE pageblocks per zone" applied. It has been
sent for consideration in -mm.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>