oprofile_map_event: Add ppc64 mappings for pagetable walks
The PPC64 PMU has events for DTLB and also for counting the cycles spent
when walking page tables. When working out what the cost in cycles of a
DTLB miss is, this event needs to be measured as the automatic means of
measuring a TLB miss calculates the cost of an ERAT miss on ppc64 instead
of the DTLB miss.

This patch adds a mapping for the necessary event.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>