oprofile_map_event: Decrease the sample period for the PPC64 events
The PPC970MP supports a minimum sample period of 10000 but the tool is
automatically mappings for a period of 50000. The value was chosen to
reduce the overhead of oprofile but the --sample-cycle-factor switch is
there to achieve the same thing. Make the default sample period as low
as is supported.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>