libhugetlbfs: Fix some small errors / warts in run_tests.py
This fixes one real bug in run_tests.py and a couple of other places
where we do things in a silly way.

The real bug is that we clearly want to use the current libhugetlbfs
binaries and libraries to run the test, so they should be added to the
front of the path, not the back as cmd_env() currently does.

In the same piece of code we use Python's `` operators for no good
reason, since the string formatting operation we're already using will
encode the bits value properly.

cmd_env() also takes an optional pagesize parameter.  The only time
it's not supplied is when it's used to invoke hugeadm, which doesn't
really suit use of cmd_env() anyway.  So, this patch makes that
parameter mandatory which simplifies the logic a bit (in fact that
test logic was also wrong, but it didn't really matter).

Finally, in bash() we use the dictionary's update() method instead of
the manual equivalent.

Signed-off-by: David Gibson <dwg@au1.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>