Don't use a shell to invoke testcases
The old run_tests.sh was replaced with run_tests.py quite a while ago,
but in some places it still shows its shell origins.  In particular,
rather than having the Python script directly execute the individual
testcase binaries, a shell is invoked to execute each one.
Environment variables are passed in as parts of the shell command line
instead of being directly provided and so forth.

That approach is a bit aesthetically displeasing, plus it slows down
operation of the testsuite.  The slowdown is usually barely
measureable, however on really slow machines (such as simulators) or
on kernels with lots of debugging added, the extra context switches
can be an annoying delay.

Therefore, this patch reworks the test script to directly invoke each
testcase executable, without invoking an intermediate shell.  This is
done through a new run_test_prog() function which combines the
existing bash() and cmd_env() functions.  The bash() which runs
through a shell is still used in a few places where we invoke whole
shell pipelines, but it's no longer used on each testcase.

Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>