Make tests for SPECIAL in linker scripts safe against cross-compiles
Currently, run_tests.py invokes gcc to obtain the standard linker
script, in order to tell if the old-style linker-script based relink
testcases can work.  But this a) assumes that a compiler and linker is
available and b) assumes that it has the same linker script as when
the testsuite was compiled.

If the testsuite is compiled on one machine, then copied to another
for execution, either of these could be false.  This is a useful
common case if the machine we're testing on is highly constrained
(e.g. an embedded machine, or a simulator).  What we really want is
the linker script of the toolchain which was used to build the
testsuite.

Therefore this patch removes the test-time invocation of gcc.
Instead, gcc is invoked to dump the linker script to a file when the
testsuite is built, then run_tests.py reads this file back in and
checks it for the appropriate linker magic.

Signed-off-by: David Gibson <dwg@au1.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>