Suppress ld.hugetlbfs warnings when building testsuite
Current versions of the ld.hugetlbfs wrapper script rightly warn if
you use the old, broken linker-script based method of building a
hugepage binary.  However, we still have to build things with this
mode to test them.  It's kind of nasty having a dozen or so warnings
spewing out on a perfectly full and correct build of the testsuite.

Therefore, this patch provides an environment variable to suppress
those warnings, and uses it from the testsuite makefile to get rid of
them in this case.

Signed-off-by: David Gibson <dwg@au1.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>