Fix for bitrot in linkhuge_nofd
The linkhuge_nofd testcase no longer tests what it's supposed to -
that the elflink code will correctly fall back to normal pages if
hugepages are not available.

The testcase checked this path by overriding the
hugetlbfs_unlinked_fd() function with one that always fails.  However
newer revisions of the elflink code don't use the
hugetlbfs_unlinked_fd() function, they use the lower level
hugetlbfs_unlinked_fd_for_size().

This patch overrides the correct function and make the testcase do
what it's supposed to.  Currently the test is effectively a duplicate
of the main linkhuge test.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>