hugeadm: Handle overflows when setting shmmax
When using --set-recommended-shmmax, it attempts to set the value based
on the maximum number of huge pages that can be allocated. In the event
the total number of huge pages exceeds the size of the virtual address
space, this counter can overflow and shmmax is set to a value that makes
no sense. This patch catches when this situation occurs and truncates the
recommended shmmax based on a unsigned long.

Tested on a 32-bit X86 machine.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>