Add paranoid checking of hugepage pool sizes
This patch adds code to the run_tests.py script to check the state of
all the hugepage pools before and after each testcase.  When turned on
(with the new '-c' option to run_tests.py), it will report an error if
the pool sizes are different after the test than before.

Obviously this can report bogus errors if other programs are using
hugepages on the system, or if dynamic pools are enabled.  Nonetheless
it can be useful it can be useful for detecting and tracking down
subtle kernel bugs which result in misaccounting of the pool state
(I've used it to debug such problems in a new hugepage port).

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>