Don't try to test multiple pagesizes for SysV shm
Currently, run_tests.py attempts to run each test case for each
available page size and word size.  It does this by setting the
HUGETLB_DEFAULT_PAGE_SIZE environment variable to instruct the
testcase's instance of libhugetlbfs to use the specified pagesize.

However some of the testcases use shmget() with the SHM_HUGETLB flag
to obtain hugepage backed memory, rather than mmap() or the
libhugetlbfs interfaces.  This means of obtaining hugepage memory has
no way to specify the pagesize, and will always use the system default
huge page size.

Therefore, running the test multiple times when it won't actually use
a different pagesize is silly.  Furthermore in some of the cases,
using the non-default sizes results in test failures (because we don't
allocate as much memory as we expect) which have to be suppressed with
a wrapper script which turns them into expected failures.

This patch, therefore, alters run_tests.py to run the shm only tests
with only the default page size.  This is accomplished with a new
helper function 'do_shm_test', which also folds in the setting and
restoring of the shmmax limit for added convenience.  The no longer
necessary wrappers for shm-fork and shm-getraw are removed.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>