Add scripts to measure cost of tlb miss in cycles V2
The scripts added by this patch are used to measure the cost of a tlb
miss on the running machine.  They rely on the administrator obtaining
the source for calibrator from:
http://homepages.cwi.nl/~manegold/Calibrator/v0.9e/calibrator.c
 compiling it and
stream from:
http://www.cs.virginia.edu/stream/FTP/Code/stream.c compiling it with
special options and putting them either in the search path or the same
directory as tlbmiss_cost.sh.

V2 of this patch makes several changes:
1.  It includes the entire stack from Mel Gorman that updated the
method that uses stream and oprofile as well as adding the ability
to fetch and build the stream and calibrator programs.

2. It moves the script into the contrib sub directory because the script
has dependencies not met by libhugetlbfs.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>