libhugetlbfs huge page setup helper
A common complaint we (Red Hat) get from customers is that setting up
huge pages for use with their java or database applications is complex.
While hugeadm makes this easier than it has been in the past, some
customers want brain-dead simple, fire and forget setup. The attached
patch adds a python script which attempts to implement that, asking the
user only three questions (how much memory and what users and group
should be able to use it), which should get them up and going.

Signed-off-by: Jarod Wilson <jarod@redhat.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>