Add quiet option to hugectl
By default the libhugetlbfs verbosity level is set to display messages
that are WARNINGS or ERRORS.  hugectl has the ability to make the library
more verbose, but not less without knowing the exact level to pass to
the --verbose option.  This patch addes a -q which will drop the verbosity
level by 1.

Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>