add --vmlinux option to tlbmiss_cost.sh
tlbmiss_cost.sh does not currently take vmlinux as a parameter, and calls
oprofile_start.sh, which uses the location of vmlinux.  This patch adds
an optional --vmlinux option to the script, and passes this value to calls
of oprofile_start.sh.  It also changes the call to the tlbmiss_cost.sh from
cpupcstat to pass the vmlinux value used in cpupcstat.

Signed-off-by: Avantika Mathur <mathur@us.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>