Allow MAP_NORESERVE to be used for mappings
Since 2.6.27-rc1, the kernel makes reservations for mappings at mmap()
time. This guarantees that the process that successfully calls mmap()
will successfully fault all pages within that region. This is nice
reliable behaviour but it can be the case the program wants to create a
very large sparse mapping. In this case, mmap() will fail even if the
program knows the huge pages are available.

This patch introduces a --no-reserve switch that uses MAP_NORESERVE.
mmap() will always succeed but the fault might not. Unfortunately, on
older kernels, use of MAP_NORESERVE can trigger the OOM killer. Hence,
this patch also checks the kernel version and only allows use of
MAP_NORESERVE if it's safe to do so.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Acked-by: Eric B Munson <ebmunson@us.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>