Fix bug in gethugepagesizes testcase
The gethugepagesizes testcase tests the calls to retreieve the
available page sizes against faked up sysfs contents.  This includes
testing that the gethugepagesizes() call doesn't return more entries
than the maximum it's given (which might clobber memory).

However, it does this incorrectly: specifically, validate_sizes()
checks that the n returned entries exactly match the first n expected
entries, even when the maxmum number of entries to return is smaller
than the total number of pagesizes available.  gethugepagesizes() is
not guaranteed to return the pagesizes in any particular order (based
as it is on getting a directory listing), so this test will generate
bogus failures.

In fact, for the case where the given maximum is smaller than the
total number of pagesizes, the test should accept as a pass any subset
of the expected pagesizes in the returned results.  This patch makes
the correction.

Signed-off-by: David Gibson <dwg@au1.ibm.com>
Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>