Fix pool adjust when using negative delta
hugeadm has always accepted deltas for adjusting huge page pool sizes, but
when support was added for 16G page sizes an unsigned value was used.  This
causes problems when a negative delta is used, instead of shrinking the
pool, the pool size is set to a very large number, usually crippling the
machine as the kernel attempts to use all of the memory as huge pages.
This patch changes the value to a signed long long which is still large
enough to hold a 16G page size.

Signed-off-by: Eric B Munson <emunson@mgebm.net>