Fix return value tests for hugetlbfs_test_feature
The function hugetlbfs_test_feature currently returns 0 if the
requested feature is not present, a positive integer if the
feature is present, and -ENOSYS if the requested feature is
invalid.  The way that hugetlbfs_test_feature is currently used is
any non-zero return value means the requested feature is present.
This patch corrects this usage to use only positive integer return
values to mean a feature is present.

Signed-off-by: Eric B Munson <emunson@mgebm.net>