Read /proc/self/smaps to test page size of mapping
When MAP_HUGETLB is used, the standard test for a mapping being backed by huge
pages fails because the mapping will not be on hugetlbfs.  Instead of testing
the filesystem backing the mapping, this patch adds a check of reported MMU
page size from /proc/self/smaps.  get_mapping_page_size returns the page
size that is being used for the specified mapping in bytes.

Signed-off-by: Eric B Munson <emunson@mgebm.net>
Acked-by: Mel Gorman <mel@csn.ul.ie>