shm.c: Allow static linking with shmget()
libhugetlbfs redefines shmget() so that it can transparently optionally
add the hugepage flag.  Unfortunately, it uses dlsym() to to find the
next shmget() to call in the dynamic library call chain.  This does not
work on statically-linked executables.

Work around this by making all references to libdl weak, and testing
for its presence.  If libdl symbols are not not present, directly
make the shmget syscall.

Making the dl* symbols weak does not affect dynamic executables as
libdl.so is:
* Automatically added to the library dependency list when liked with
  -lhugetlbfs
* Mapped into the programs address space when used with LD_PRELOAD.

Based on a patch originally authored by Dean Luick <luick@cray.com>.

Signed-off-by: Andrew Hastings <abh@cray.com> on behalf of Cray Inc.
Signed-off-by: Eric B Munson <emunson@mgebm.net>