hugeadm.c: Don't segfault if user not in passwd file
If a user somehow managed to run hugeadm without an entry being present
for the user in the passwd file, hugeadm was segfaulting because it
did not check for failure of getpwuid().

Fix:  check for getpwuid() failure and issue an appropriate error message.

Based on a patch originally authored by Dean Luick <luick@cray.com>.

Signed-off-by: Andrew Hastings <abh@cray.com> on behalf of Cray Inc.
Signed-off-by: Eric B Munson <emunson@mgebm.net>