Allow actions to be restricted to certain programs
By default, libhugetlbfs will act on any program that it is loaded
with, either via LD_PRELOAD or by explicitly linking with -lhugetlbfs.

There are situations in which it is desirable to restrict libhugetlbfs'
actions to specific programs.  For example, some ISV applications are
wrapped in a series of scripts that invoke bash, python, and/or perl.
It is more convenient to set the environment variables related to
libhugetlbfs outside of the wrapper scripts, yet this has the unintended
and undesirable consequence of causing the script interpreters to use
and consume hugepages.  Often there is no obvious benefit to causing the
script interpreters to use hugepages, and there is a clear disadvantage:
fewer hugepages are available to the actual application.

To address this scenario, add a HUGETLB_RESTRICT_EXE environment variable.
If this variable is set, libhugetlbfs will restrict its actions to only
those programs named in HUGETLB_RESTRICT_EXE.  (If not set, libhugetlbfs
will attempt to apply the requested actions to all programs.)

For example,

    HUGETLB_RESTRICT_EXE="hpcc:long_hpcc"

will restrict libhugetlbfs' actions to programs named /home/fred/hpcc
and /bench/long_hpcc but will not affect /usr/bin/hpcc_no.

Based on a patch originally authored by Dean Luick <luick@cray.com>.

Signed-off-by: Andrew Hastings <abh@cray.com> on behalf of Cray Inc.
Signed-off-by: Eric B Munson <emunson@mgebm.net>