Check permissions when selecting mount
Currently the library will select the first hugetlbfs mount it finds for a
given page size for usage.  This can lead to frustrating failures when root
owns the first mount and a user has a later mount in the list.  This patch
makes find_mounts check if a given mount is usable by the current user.  Mounts
that are not usable will be skipped.

Signed-off-by: Eric B Munson <emunson@mgebm.net>
Cc: Andrew Lutomirski <luto@mit.edu>
Cc: David Gibson <david@gibson.dropbear.id.au>