Decrement index before closing files on error exit
If obtain_prepared_file fails for any reason, the last fd in our hugetlb segment
table will be 0 when the error clean up starts because obtain_prepared_file
does not write to the segment structure in error cases.  Because the cleanup
for loop starts at the segment where obtain_prepared_file failed it will try
to close stdout (fd 0).  This patch makes the cleanup for loop decrement the
index before starting to close file descriptors.

Reported-by: Dan Scales <scales@vmware.com>
Signed-off-by: Eric B Munson <emunson@mgebm.net>