Make custom linker scripts configurable
This patch introduces a CUSTOM_LDSCRIPTS variable in the Makefile as
preparation for enabling libhugetlbfs linking on architectures w/o providing
(deprecated) custom linker scripts. Setting it to "no" will prevent building
and running the test cases requiring custom linker scripts, and issue an
error message in ld.hugetlbfs when option --hugetlbfs-link is being used.
It is set to "yes" by default for all architectures, which means no change
to the current behaviour as far as "make" is concerned.

A new option -l is added to run_tests.py, enabling the old-style linking
tests. This means that the default for running run_tests.py directly is
now changed to skip the old-style linking tests.

Signed-off-by: Gerald Schaefer <gerald.schaefer@de.ibm.com>
Signed-off-by: Eric B Munson <emunson@mgebm.net>