Fix WARNING: Unable to verify address range
This patch fixes a warning about an empty address range that will be issued
when the offset in prepare_segment() is smaller than page_size. In that case,
check_range_empty() would be called with len == 0 (the page_size down-aligned
offset), triggering the warning. This is fixed by checking the down-aligned
offset instead of the offset itself before calling check_range_empty().

Signed-off-by: Gerald Schaefer <gerald.schaefer@de.ibm.com>
Signed-off-by: Eric B Munson <emunson@mgebm.net>