RE: [Libhugetlbfs-devel] update for clipped LD_LIBRARY_PATH
This patch resolves clipped LD_LIBRARY_PATH.  Using hugectl to enable use of hugepage for an application on a system with SLES11 SP1.  hugectl was clipping my LD_LIBRARY_PATH at 254 characters and preventing execution.  NAME_MAX on my platform is 255.  This patch allows LD_LIBRARY_PATH to be up to 4096 characters on my platform.

Signed-off-by: Daniel Kokron <Daniel.Kokron@nasa.gov>
Signed-off-by: Eric B Munson <emunson@mgebm.net>