task-size-overrun: fix problem with dynamic pagetable upgrade on s390x
The strategy to find out TASK_SIZE won't work on s390x anymore, starting
with kernel 3.9. We will dynamically increase the pagetable levels on
s390x on access beyond TASK_SIZE, effectively increasing TASK_SIZE from
2^42 to 2^53, but /proc/self/maps won't reflect this.

With the current strategy that means that find_task_size() would loop
for a very long time, from 2^42 to 2^53. To fix this, increase addr
in the loop for s390x as soon as we exceed the 2^42 limit.

Signed-off-by: Gerald Schaefer <gerald.schaefer@de.ibm.com>
Signed-off-by: Eric B Munson <emunson@mgebm.net>