tests: add corrupt-by-cow-opt test case
Current kernel (v3.10) or older has a problem in cow optimization. Cow
optimization is that if we find a page which has only 1 reference from
this process, we use it directly for private write. This applied to
anonymous page and a page from page cache until now. This cause a problem
for a page cache case, because we can remap this file by another process,
and it can see corrupt data. To detect this problem, add this test case.

This problem will be fixed on v3.12 kernel.

Signed-off-by: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Signed-off-by: Eric B Munson <emunson@mgebm.net>