Cleanup ARM ld.hugetlbfs HTLB_LINK logic
When ld.hugetlbfs is executed with --hugetlbfs-link, there is code
to check for the ARM platform and warn that this is not supported.

There is also code to check for CUSTOM_LDSCRIPTS being false and
give a similar warning.

This patch removes the ARM check as the CUSTOM_LDSCRIPTS check will
catch this.

Signed-off-by: Steve Capper <steve.capper@linaro.org>
Signed-off-by: Eric B Munson <emunson@mgebm.net>