simplify slice boundary search
init_slice_boundary() was meant to find 2 neighbour slices,
which are both free. It's not needed to have 2 completely
free slices as the test needs only couple huge pages around
slice boundary.

As part of search in 32bit version it tried to mmap at 0, which
may trigger selinux avc:  denied  { mmap_zero }.

This patch simplifies search function, so that it tries mmap 2 huge pages
(from each side) around slice boundary. If successful this slice
boundary is used for the rest of test.

Signed-off-by: Jan Stancek <jstancek@redhat.com>
Signed-off-by: Eric B Munson <emunson@mgebm.net>