ARM: mark direct_syscall as a FUNC
direct_syscall is ARM state assembler. If libhugetlbfs is compiled for
the Thumb state it does not currently change execution states when it
calls direct_syscall, leading to unpredictable behaviour.

This patch marks direct_syscall as a FUNC, which enables the ARM/Thumb
state interworking; resulting in a BLX to direct_syscall where
necessary.

Signed-off-by: Steve Capper <steve.capper@linaro.org>
Signed-off-by: Eric B Munson <emunson@mgebm.net>