libhugetlbfs: segment-based alignment restrictions exist on pseries
Commit 8d5390c4 ("libhugetlbfs: Remove segment-based alignment
restrictions") defined the CPPFLAG PPC_NO_SEGMENTS, because BookE
Freescale processors don't have this restriction. However it is defined
globally and thus is also applied to 32-bit builds on pseries hardware.
This is simply incorrect and breaks the 32-bit malloc testcases (which
fail to align the heap address correctly) on ppc64, because there is no
situation where PPC_NO_SEGMENTS is not defined.

I believe the right fix is to only define this flag when we are on a
natively 32-bit platform. With this change, the malloc testcase pass
again.

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>
Tested-by: Jan Stancek <jstancek@redhat.com>
Signed-off-by: Eric B Munson <emunson@mgebm.net>