Update README with instructions on how to submit patches
The move to Github opened up the pull-request tool that they provide.
Because not all libhugetlbfs developers have a github account it is
preferred that we continue with the tried and true method of sending
patches over the mailing list.  The basics of how to do this are covered
in the SubmittingCode document.

Signed-off-by: Eric B Munson <emunson@akamai.com>