tests/linkhuge_rw: function ptr may not refer to .text
On some ABIs function pointer may not refer to .text section.
For example on powerPC 64-bit ABI, function pointer may refer
to a call stub from .opd section.

This creates a problem for linkhuge_rw tests which run with
HUGETLB_ELFMAP=R, because test is expecting that address of
function pointer will be backed by huge pages. But because
.opd section is from RW PT_LOAD segment, this doens't happen,
since libhugetlbfs is instructed to map only R segments via
HUGETLB_ELFMAP=R.

This patch is replacing use of function pointer with address
returned by gcc's __builtin_return_address(), that is called
by the function itself. This should provide an address that
is from an actual code, residing in .text section.

Signed-off-by: Jan Stancek <jstancek@redhat.com>
Cc: Adam Litke <agl@us.ibm.com>
Cc: Eric B Munson <emunson@mgebm.net>
Cc: Petr Holasek <pholasek@redhat.com>
Signed-off-by: Eric B Munson <emunson@mgebm.net>