huge_page_setup_helper: do not assume default huge page size is first
The helper script currently implicitly assumes that `hugeadm
--pool-list` will emit the default size first, as it assumes every line
has 5 fields in it. But only the default field does, and if `hugeadm
--pool-list` were to output:

hugeadm --pool-list
      Size  Minimum  Current  Maximum  Default
   1048576        0        0        0
  16777216        0        0        0        *
17179869184        0        0        0

we receive an error from the script:
 # ./huge_page_setup_helper.py
Traceback (most recent call last):
  File "./huge_page_setup_helper.py", line 51, in <module>
    if line.split()[4] == '*':
IndexError: list index out of range

Just check for the '*' character to determine the default line.

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>
Signed-off-by: Eric B Munson <emunson@mgebm.net>