Fix behavior while shrinking.
Adjust mapsize as we're unmapping pages.  Do not lie to glibc about
shrinking by less than a page.  It's unncessary because we are not
giving back any memory to the kernel, but also it forces us to zero
out this memory because morecore() assumes by default that "new"
memory is already zero'd.

Signed-off-by: Guillaume Morin <guillaume@morinfr.org>