tests: Detect not present tests in run_tests.py
If a test is not present in the tests folder, it will
be accounted as "Killed by signal". This is quite generic
as the test could have failed by any signal.

This patch returns -errno when an OSError occurs.
It checkes if the test file wasn't present in the file system
and increments a new result_type which counts excatly this case.

Signed-off-by: Matthias Brugger <mbrugger@suse.com>
Signed-off-by: Eric B Munson <emunson@mgebm.net>