tests/heapshrink: disable malloc per-thread cache for accuracy
After upstream commit: (glibc-2.25.90-688-gd5c3fafc43) glibc has a new
per-thread caching mechanism that will NOT allow this test to
successfully measure if heap has shrunk, or not, due to the fact that
heap won't have its sized reduced right away, as the tests expects.

Only way to test heap shrinkness in an accurate way is to disable this
per-thread cache. In order to disable it you need to have an environment
tunable GLIBC in place:

GLIBC_TUNABLES=glibc.malloc.tcache_count=0

Unfortunately it requires to be set before program is loaded, since it
is evaluated during malloc() initialization, and there is no way to
re-initialize malloc() from the program context (not even using
constructor functions).

This commit sets GLIBC_TUNABLES environment variable for the heapshrink
tests in order to make the test to succeed in environments with newer
GLIBC. Older environments won't suffer any consequences since the
environment tunable will just be ignored.

Link: https://bugs.linaro.org/show_bug.cgi?id=3950
Signed-off-by: Rafael David Tinoco <rafael.tinoco@linaro.org>
Signed-off-by: Eric B Munson <emunson@mgebm.net>