tests: Explicitly decode subprocess output
The output we get from subprocesses is logically a sequence of bytes, but
we want to treat it as Python strings, which means decoding it into Unicode
based on some encoding.

In Python2 we can get away with skipping that step, but in Python3 we won't
be able to.  So, to get ready, add an explicit decode() step, using the
system default encoding (probably UTF-8 in most cases).

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>