tests: Improve TASK_SIZE detection in task-size-overrun
task-size-overrun is designed to test kernel behaviour in some edge cases
involving making a hugepage mapping right near the address space limits.
In order to do that, it needs to know the TASK_SIZE of the kernel it's
running on.

Currently it does that with a linear search from the last extant mapping.
But with kernels supporting a very large address space that can take
prohibitively long.  We've had problems with that before, resulting in some
hacks to skip a large chunk of address space.

Those hacks are dependent on platform, though, which is ugly and fragile.
Case in point, recent powerpc kernels now support a 4PiB address space,
so the logic we have there is insufficient to finish the search in
reasonable time.

To handle this in a more robust way, this replaces the linear search with
a binary search between the last extant mapping and (2^wordsize).

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Eric B Munson <eric@munsonfam.org>