Wait child with os.wait()
os.popen() is an async method, it fork() child and exec() in child with
the arg command. If child is slow enough relative to the parent process,
this could lead to unexpected result where parent need to sync with child.

During our test, we find python3 is faster than python2, after coverting
to python3, 'groupadd' usually doesn't finish when the followed iter on
groups starts, we would get '-1' as the groupid and lead to error.

To reproduce it with python3:
/root/rpmbuild/BUILD/libhugetlbfs-2.21/huge_page_setup_helper.py <<EOF
128
hugepages
hugepages root
EOF

...
hugeadm:ERROR: Invalid group specification (-1)
...

Signed-off-by: Chunyu Hu <chuhu@redhat.com>
Signed-off-by: Eric B Munson <eric@munsonfam.org>