Add missing '\n' newline characters to WARNING messages
A number of WARNING messages were missing trailing '\n' newline
characters, which can result in user output being emitted on the same
line as WARNING messages.  Based on other WARNING messages, the intent
appears to be to include a trailing '\n' newline, so add them where
missing.  This issue was observed for the "Hugepage size %li
unavailable" WARNING message.